Split off core/artifact_tree.py
